app/requirements.txt
pyyaml
dnspython
httpx[http2]
h2
aioquic
flask